move rotation methods to modular component @WXR-014
